requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.23.0
attrs==20.3.0
autopep8==1.5.4
base58==2.1.0
beautifulsoup4==4.10.0
boto3==1.17.86
botocore==1.20.86
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
discord.py @ git+https://github.com/Rapptz/discord.py@9648986028a9721e1ba733750eb8c30e163dcb2c
ECPy==1.2.5
et-xmlfile==1.1.0
humanize==3.9.0
idna==2.10
jmespath==0.10.0
lxml==4.6.3
multidict==5.1.0
numpy==1.21.0
openai==0.19.0
openpyxl==3.0.10
pandas==1.4.2
pandas-stubs==1.2.0.61
plyer==2.0.0
psutil==5.8.0
pycodestyle==2.6.0
pycparser==2.20
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.1
python-socks==1.2.3
pytz==2021.1
PyYAML==6.0
requests==2.25.1
requests-tor==0.5
s3transfer==0.4.2
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
stem==1.8.0
tabulate==0.8.9
tbselenium==0.5.3
toml==0.10.2
torpy==1.1.5
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarg==0.1.9
yarl==1.6.3
